+++ crates/kct_jsonnet/src/error.rs
@@ -1,0 +1,21 @@
+use thiserror::Error;
+
+#[derive(Error, Debug, PartialEq, Eq)]
+pub enum Error {
+	#[error("No target to compile")]
+	NoTarget,
+	#[error("No validation provided")]
+	NoValidator,
+	#[error("No input was provided")]
+	NoInput,
+	#[error("The input provided doesn't match the schema: {0}")]
+	InvalidInput(String),
+	#[error("An error happened while rendering your templates: {0}")]
+	RenderIssue(String),
+	#[error("Your template couldn't be parsed as JSON")]
+	InvalidOutput,
+	#[error("{0}")]
+	Wrapped(String),
+}
+
+pub type Result<T> = std::result::Result<T, Error>;